pyproject.toml
[tool.poetry]
name = "test_sanic"
version = "0.1.0"
description = ""
authors = ["vladislav <vladislav.shirobokov@nlogic.ai>"]

[tool.poetry.dependencies]
python = "^3.10"
sanic = "^22.9.0"
sanic-ext = "^22.9.0"
pydantic = "^1.10.2"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.2"}
python-dotenv = "^0.21.0"
aerich = "^0.7.1"
pycrypto = "^2.6.1"
pyjwt = "^2.6.0"
aiohttp = "^3.8.3"
sanic-testing = "^22.9.0"
pytest-asyncio = "^0.20.1"

[tool.poetry.dev-dependencies]
mypy = "^0.982"
flake8 = "^5.0.4"


[tool.aerich]
tortoise_orm = "config.setting.TORTOISE_CONFIG"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
